Implement permissions + services

+++ app/Http/Controllers/TicketController.php
@@ -7,21 +7,25 @@ use App\Http\Requests\Ticket\TicketCreateRequest;
 use App\Http\Requests\Ticket\TicketDeleteRequest;
 use App\Http\Requests\Ticket\TicketUpdateRequest;
 use App\Http\Responses\Response;
+use App\Http\Services\TicketService;
 use App\Models\Event;
 use App\Models\Ticket;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class TicketController extends Controller
 {
+    /**
+     * Returns all tickets
+     */
     public function index(): JsonResponse
     {
-        $tickets = Ticket::all();
+        $tickets = TicketService::getAll();
         return Response::success($tickets);
     }
 
     /**
-     * Returns a single Event
+     * Returns a single ticket
      */
     public function show(Ticket $ticket): JsonResponse
     {
@@ -29,17 +33,16 @@ class TicketController extends Controller
     }
 
     /**
-     * Creates a event and saves it to the database
+     * Creates a ticket and saves it to the database
      */
     public function store(TicketCreateRequest $request): JsonResponse
     {
-        $ticketDTO = TicketDTO::arrayFromRequest($request);
-        $ticket = Ticket::create($ticketDTO);
+        $ticket = TicketService::create($request);
         return Response::success($ticket, 201);
     }
 
     /**
-     * Update a event and saves it to the database
+     * Update a ticket and saves it to the database
      */
     public function update(TicketUpdateRequest $request): JsonResponse
     {
@@ -48,21 +51,30 @@ class TicketController extends Controller
         $ticket = Ticket::where('user_id', $ticketDTO['user_id'])
             ->where('event_id', $ticketDTO['event_id'])
             ->firstOrFail();
+        
+        if (!Gate::allows('manage-event', $ticket)) {
+            return Response::error((object) ['error' => 'Unauthorized.'], 403);
+        }
             
         $ticket->update($ticketDTO);
         return Response::success($ticket);
     }
     
     /**
-     * Deletes a event from the database
+     * Deletes a ticket from the database
      */
-    public function destroy(TicketDeleteRequest $request): JsonResponse
+    public function destroy(Event $event, TicketDeleteRequest $request): JsonResponse
     {
         $ticketDTO = TicketDTO::arrayFromRequest($request);
 
         $ticket = Ticket::where('user_id', $ticketDTO['user_id'])
-            ->where('event_id', $ticketDTO['event_id'])
+            ->where('event_id', $event->id)
             ->firstOrFail();
+        
+
+        if (!Gate::allows('manage-event', $ticket)) {
+            return Response::error((object) ['error' => 'Unauthorized.'], 403);
+        }
 
         $ticket->delete();
         return Response::success((object) [], 204);
